refactor: getClothes 타입 Clothes 폴더로 이동
