+++ src/type/clothes/getClothes.res.ts
